+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## LethalLib [1.0.2]
+
+### Added
+
+- Validation to scrap and shop items registered with invalid weight values (above 4 and under 1).
+
 ## LethalLib [1.0.1]
 
 ### Added